Added sprint 6 rake task and VCR cassettes for OSC Physics
