+++ Gemfile
@@ -136,6 +136,12 @@ group :development, :test do
   # Lorem Ipsum
   gem 'faker'
 
+  # Stubs HTTP requests
+  gem 'webmock'
+
+  # Records HTTP requests
+  gem 'vcr'
+
   # Automated security checks
   gem 'brakeman'
