+++ app/models/task_step.rb
@@ -6,6 +6,7 @@ class TaskStep < ActiveRecord::Base
   validates :tasked, presence: true
   validates :tasked_id, uniqueness: { scope: :tasked_type }
   validates :title, presence: true
+  validates :content, presence: true
 
   def complete
     self.completed_at ||= Time.now
