+++ app/routines/import/book.rb
@@ -4,16 +4,17 @@ module Import
     lev_routine
 
     uses_routine Import::CnxResource,
-                 as: :cnx_import, translations: { outputs: { type: :verbatim } }
+                 as: :cnx_import,
+                 translations: { outputs: { type: :verbatim } }
 
-    uses_routine Import::Page, as: :import
+    uses_routine Import::Page, as: :page_import
 
     protected
 
     # Recursively imports items in a CNX collection into the given book
     def import_collection(parent_book, hash, options = {})
       book = ::Book.create(parent_book: parent_book,
-                         title: hash['title'] || '')
+                           title: hash['title'] || '')
 
       parent_book.child_books << book unless parent_book.nil?
 
@@ -21,7 +22,8 @@ module Import
         if item['id'] == 'subcol'
           import_collection(book, item, options)
         else
-          run(:import, item['id'], book, options.merge(title: item['title']))
+          run(:page_import, item['id'], book, 
+                            options.merge(title: item['title']))
         end
       end
 
@@ -31,7 +33,7 @@ module Import
     # Imports and saves a CNX book as a Book
     # Returns the Book object, Resource object and collection JSON as a hash
     def exec(id, options = {})
-      run(:cnx_import, id, options)
+      run(:cnx_import, id, options.merge(book: true))
 
       outputs[:book] = import_collection(nil, outputs[:hash]['tree'], options)
       outputs[:book].resource = outputs[:resource]
