+++ lib/user_mapper.rb
@@ -3,7 +3,7 @@ module UserMapper
   def self.account_to_user(account)
     return User.anonymous if account.is_anonymous?
 
-    user = User.where{account_id == account.id}.first
+    user = User.where(account_id: account.id).first
     return user if user.present?
 
     outcome = CreateUser.call(account)
