+++ spec/factories/assistants.rb
@@ -3,7 +3,7 @@ class EmptyClass; end
 FactoryGirl.define do
   factory :assistant do
     name { Faker::Name.name }
-    code_class_name "Dummy"
+    code_class_name "DummyAssistant"
     task_plan_type "dummy"
   end
 end
