--- spec/mocks/assistants/dummy.rb
@@ -1,10 +1,0 @@
-module Assistants
-  class Dummy
-    def self.schema
-      "{}"
-    end
-
-    def self.distribute_tasks(task_plan:, taskees:)
-    end
-  end
-end
