+++ spec/mocks/assistants/dummy_assistant.rb
@@ -1,0 +1,8 @@
+class DummyAssistant
+  def self.schema
+    "{}"
+  end
+
+  def self.distribute_tasks(task_plan:, taskees:)
+  end
+end
