+++ spec/models/task_step_spec.rb
@@ -8,6 +8,8 @@ RSpec.describe TaskStep, :type => :model do
 
   it { is_expected.to validate_presence_of(:task) }
   it { is_expected.to validate_presence_of(:tasked) }
+  it { is_expected.to validate_presence_of(:title) }
+  it { is_expected.to validate_presence_of(:content) }
 
   it "requires tasked to be unique" do
     expect(task_step).to be_valid
